split out built-in type
